+++ src/modules/feature-modules/accessor/pages/innovation/needs-assessment-overview/needs-assessment-overview.component.html
@@ -7,7 +7,7 @@
 
         <div class="app-pane__side-bar">
           <p class="nhsuk-heading-xs">{{ assessment?.finishedAt | date: ('app.date_formats.long_date' | translate) }}</p>
-          <p class="nhsuk-hint nhsuk-u-font-weight-bold nhsuk-u-margin-bottom-0">{{ assessment?.assignToName }}</p>
+          <p class="nhsuk-hint nhsuk-u-font-weight-bold nhsuk-u-margin-bottom-0">{{ assessment?.assignTo?.name }}</p>
           <p class="nhsuk-hint">Needs assessment team</p>
         </div>
 
@@ -61,7 +61,7 @@
               <h2 class="nhsuk-heading-l nhsuk-u-margin-bottom-3">Suggested organisations</h2>
               <div *ngFor="let organisation of suggestedOrganisations">
                 <p class="nhsuk-u-margin-bottom-1"> {{ organisation.name }} </p>
-                <div *ngFor="let unit of organisation?.organisationUnits">
+                <div *ngFor="let unit of organisation?.units">
                   <ul *ngIf="organisation.name !== unit.name" class="nhsuk-list nhsuk-u-margin-left-3 nhsuk-u-margin-bottom-0 border-left-inset-neutral">
                     <li class="d-flex nhsuk-u-margin-0 nhsuk-u-padding-left-3">
                       <span class="app-task-list__task-name"> {{ unit.name }} </span>
