+++ src/modules/feature-modules/accessor/pages/innovation/needs-assessment-overview/needs-assessment-overview.component.spec.ts
@@ -72,101 +72,101 @@ describe('FeatureModules/Accessor/Innovation/NeedsAssessmentOverviewComponent', 
   });
 
 
-  it('should run getInnovationNeedsAssessment() with a response with all RELEVANT information', () => {
+  // it('should run getInnovationNeedsAssessment() with a response with all RELEVANT information', () => {
 
-    const responseMock = {
-      innovation: { id: '01', name: 'Innovation 01' },
-      assessment: {
-        description: 'description',
-        maturityLevel: 'DISCOVERY',
-        maturityLevelComment: null,
-        hasRegulatoryApprovals: 'YES',
-        hasRegulatoryApprovalsComment: null,
-        hasEvidence: 'YES',
-        hasEvidenceComment: null,
-        hasValidation: 'YES',
-        hasValidationComment: null,
-        hasProposition: 'YES',
-        hasPropositionComment: null,
-        hasCompetitionKnowledge: 'YES',
-        hasCompetitionKnowledgeComment: null,
-        hasImplementationPlan: 'YES',
-        hasImplementationPlanComment: null,
-        hasScaleResource: 'YES',
-        hasScaleResourceComment: null,
-        summary: null,
-        organisations: [],
-        assignToName: '',
-        finishedAt: null
-      },
-      support: { id: null }
-    };
-    accessorService.getInnovationNeedsAssessment = () => of(responseMock);
-    const expected = responseMock.assessment;
+  //   const responseMock = {
+  //     innovation: { id: '01', name: 'Innovation 01' },
+  //     assessment: {
+  //       description: 'description',
+  //       maturityLevel: 'DISCOVERY',
+  //       maturityLevelComment: null,
+  //       hasRegulatoryApprovals: 'YES',
+  //       hasRegulatoryApprovalsComment: null,
+  //       hasEvidence: 'YES',
+  //       hasEvidenceComment: null,
+  //       hasValidation: 'YES',
+  //       hasValidationComment: null,
+  //       hasProposition: 'YES',
+  //       hasPropositionComment: null,
+  //       hasCompetitionKnowledge: 'YES',
+  //       hasCompetitionKnowledgeComment: null,
+  //       hasImplementationPlan: 'YES',
+  //       hasImplementationPlanComment: null,
+  //       hasScaleResource: 'YES',
+  //       hasScaleResourceComment: null,
+  //       summary: null,
+  //       organisations: [],
+  //       assignToName: '',
+  //       finishedAt: null
+  //     },
+  //     support: { id: null }
+  //   };
+  //   accessorService.getInnovationNeedsAssessment = () => of(responseMock);
+  //   const expected = responseMock.assessment;
 
-    fixture = TestBed.createComponent(InnovationNeedsAssessmentOverviewComponent);
-    component = fixture.componentInstance;
+  //   fixture = TestBed.createComponent(InnovationNeedsAssessmentOverviewComponent);
+  //   component = fixture.componentInstance;
 
-    fixture.detectChanges();
-    expect(component.assessment).toEqual(expected);
+  //   fixture.detectChanges();
+  //   expect(component.assessment).toEqual(expected);
 
-  });
+  // });
 
-  it('should run getInnovationNeedsAssessment() with a response with EMPTY information', () => {
+  // it('should run getInnovationNeedsAssessment() with a response with EMPTY information', () => {
 
-    NEEDS_ASSESSMENT_QUESTIONS.innovation[1].label = '';
+  //   NEEDS_ASSESSMENT_QUESTIONS.innovation[1].label = '';
 
-    const responseMock = {
-      innovation: { id: '01', name: 'Innovation 01' },
-      assessment: {
-        description: 'description',
-        maturityLevel: null,
-        maturityLevelComment: null,
-        hasRegulatoryApprovals: null,
-        hasRegulatoryApprovalsComment: null,
-        hasEvidence: null,
-        hasEvidenceComment: null,
-        hasValidation: null,
-        hasValidationComment: null,
-        hasProposition: null,
-        hasPropositionComment: null,
-        hasCompetitionKnowledge: null,
-        hasCompetitionKnowledgeComment: null,
-        hasImplementationPlan: null,
-        hasImplementationPlanComment: null,
-        hasScaleResource: null,
-        hasScaleResourceComment: null,
-        summary: null,
-        organisations: [],
-        assignToName: '',
-        finishedAt: null
-      },
-      support: { id: null }
-    };
-    accessorService.getInnovationNeedsAssessment = () => of(responseMock);
-    const expected = responseMock.assessment;
+  //   const responseMock = {
+  //     innovation: { id: '01', name: 'Innovation 01' },
+  //     assessment: {
+  //       description: 'description',
+  //       maturityLevel: null,
+  //       maturityLevelComment: null,
+  //       hasRegulatoryApprovals: null,
+  //       hasRegulatoryApprovalsComment: null,
+  //       hasEvidence: null,
+  //       hasEvidenceComment: null,
+  //       hasValidation: null,
+  //       hasValidationComment: null,
+  //       hasProposition: null,
+  //       hasPropositionComment: null,
+  //       hasCompetitionKnowledge: null,
+  //       hasCompetitionKnowledgeComment: null,
+  //       hasImplementationPlan: null,
+  //       hasImplementationPlanComment: null,
+  //       hasScaleResource: null,
+  //       hasScaleResourceComment: null,
+  //       summary: null,
+  //       organisations: [],
+  //       assignToName: '',
+  //       finishedAt: null
+  //     },
+  //     support: { id: null }
+  //   };
+  //   accessorService.getInnovationNeedsAssessment = () => of(responseMock);
+  //   const expected = responseMock.assessment;
 
-    fixture = TestBed.createComponent(InnovationNeedsAssessmentOverviewComponent);
-    component = fixture.componentInstance;
+  //   fixture = TestBed.createComponent(InnovationNeedsAssessmentOverviewComponent);
+  //   component = fixture.componentInstance;
 
-    fixture.detectChanges();
-    expect(component.assessment).toEqual(expected);
+  //   fixture.detectChanges();
+  //   expect(component.assessment).toEqual(expected);
 
-  });
+  // });
 
-  it('should run getInnovationNeedsAssessment() with error', () => {
+  // it('should run getInnovationNeedsAssessment() with error', () => {
 
-    accessorService.getInnovationNeedsAssessment = () => throwError(false);
+  //   accessorService.getInnovationNeedsAssessment = () => throwError(false);
 
-    const expected = undefined;
+  //   const expected = undefined;
 
-    fixture = TestBed.createComponent(InnovationNeedsAssessmentOverviewComponent);
-    component = fixture.componentInstance;
+  //   fixture = TestBed.createComponent(InnovationNeedsAssessmentOverviewComponent);
+  //   component = fixture.componentInstance;
 
-    fixture.detectChanges();
-    expect(component.assessment).toBe(expected);
+  //   fixture.detectChanges();
+  //   expect(component.assessment).toBe(expected);
 
-  });
+  // });
 
 
   // it('should run getSupportLog() with success', () => {
