+++ src/modules/feature-modules/accessor/services/accessor.service.ts
@@ -164,42 +164,6 @@ export type getActionsListEndpointInDTO = {
 };
 export type getActionsListEndpointOutDTO = { count: number, data: (getActionsListEndpointInDTO['data'][0] & { name: string })[] };
 
-export type getInnovationNeedsAssessmentEndpointInDTO = {
-  id: string;
-  innovation: { id: string; name: string; };
-  description: null | string;
-  maturityLevel: null | string;
-  maturityLevelComment: null | string;
-  hasRegulatoryApprovals: null | string;
-  hasRegulatoryApprovalsComment: null | string;
-  hasEvidence: null | string;
-  hasEvidenceComment: null | string;
-  hasValidation: null | string;
-  hasValidationComment: null | string;
-  hasProposition: null | string;
-  hasPropositionComment: null | string;
-  hasCompetitionKnowledge: null | string;
-  hasCompetitionKnowledgeComment: null | string;
-  hasImplementationPlan: null | string;
-  hasImplementationPlanComment: null | string;
-  hasScaleResource: null | string;
-  hasScaleResourceComment: null | string;
-  summary: null | string;
-  organisations: {
-    id: string; name: string; acronym: null | string;
-    organisationUnits: { id: string; name: string; acronym: null | string; }[];
-  }[];
-  assignToName: string;
-  finishedAt: null | string;
-  support: { id: null | string; }
-};
-
-export type getInnovationNeedsAssessmentEndpointOutDTO = {
-  innovation: { id: string; name: string; };
-  assessment: Omit<getInnovationNeedsAssessmentEndpointInDTO, 'id' | 'innovation' | 'support'>,
-  support: getInnovationNeedsAssessmentEndpointInDTO['support']
-};
-
 
 export type getInnovationSupportsDTO = {
   id: string;
@@ -430,44 +394,6 @@ export class AccessorService extends CoreService {
     );
 
   }
-
-  getInnovationNeedsAssessment(innovationId: string, assessmentId: string): Observable<getInnovationNeedsAssessmentEndpointOutDTO> {
-
-    const url = new UrlModel(this.API_URL).addPath('accessors/:userId/innovations/:innovationId/assessments/:assessmentId').setPathParams({ userId: this.stores.authentication.getUserId(), innovationId, assessmentId });
-    return this.http.get<getInnovationNeedsAssessmentEndpointInDTO>(url.buildUrl()).pipe(
-      take(1),
-      map(response => ({
-        innovation: response.innovation,
-        assessment: {
-          description: response.description,
-          maturityLevel: response.maturityLevel,
-          maturityLevelComment: response.maturityLevelComment,
-          hasRegulatoryApprovals: response.hasRegulatoryApprovals,
-          hasRegulatoryApprovalsComment: response.hasRegulatoryApprovalsComment,
-          hasEvidence: response.hasEvidence,
-          hasEvidenceComment: response.hasEvidenceComment,
-          hasValidation: response.hasValidation,
-          hasValidationComment: response.hasValidationComment,
-          hasProposition: response.hasProposition,
-          hasPropositionComment: response.hasPropositionComment,
-          hasCompetitionKnowledge: response.hasCompetitionKnowledge,
-          hasCompetitionKnowledgeComment: response.hasCompetitionKnowledgeComment,
-          hasImplementationPlan: response.hasImplementationPlan,
-          hasImplementationPlanComment: response.hasImplementationPlanComment,
-          hasScaleResource: response.hasScaleResource,
-          hasScaleResourceComment: response.hasScaleResourceComment,
-          summary: response.summary,
-          organisations: response.organisations,
-          assignToName: response.assignToName,
-          finishedAt: response.finishedAt,
-        },
-        support: response.support
-      })
-      )
-    );
-
-  }
-
 
   getInnovationSupportInfo(innovationId: string, supportId: string): Observable<{ id: string, status: keyof typeof INNOVATION_SUPPORT_STATUS, accessors: { id: string, name: string }[] }> {
