+++ src/modules/feature-modules/innovator/pages/innovation/needs-assessment-overview/needs-assessment-overview.component.html
@@ -8,12 +8,12 @@
         <div class="app-pane__side-bar">
 
           <p class="nhsuk-heading-xs">Updated: {{ assessment?.updatedAt | date: ('app.date_formats.long_date' | translate) }}</p>
-          <p class="nhsuk-hint nhsuk-u-font-weight-bold nhsuk-u-margin-bottom-0">{{ assessment?.updatedBy }}</p>
+          <p class="nhsuk-hint nhsuk-u-font-weight-bold nhsuk-u-margin-bottom-0">{{ assessment?.updatedBy?.name }}</p>
           <p class="nhsuk-hint">Needs assessment team</p>
 
           <div class="nhsuk-u-padding-top-8">
             <p class="nhsuk-heading-xs">Created: {{ assessment?.finishedAt | date: ('app.date_formats.long_date' | translate) }}</p>
-            <p class="nhsuk-hint nhsuk-u-font-weight-bold nhsuk-u-margin-bottom-0">{{ assessment?.assignToName }}</p>
+            <p class="nhsuk-hint nhsuk-u-font-weight-bold nhsuk-u-margin-bottom-0">{{ assessment?.assignTo?.name }}</p>
             <p class="nhsuk-hint">Needs assessment team</p>
           </div>
         </div>
@@ -66,9 +66,9 @@
               <p class="text-pre-wrap">{{ assessment?.summary }}</p>
 
               <h2 class="nhsuk-heading-l nhsuk-u-margin-bottom-3">Suggested organisations</h2>
-              <div *ngFor="let organisation of assessment?.organisations">
+              <div *ngFor="let organisation of assessment?.suggestedOrganisations">
                 <p class="nhsuk-u-margin-bottom-1"> {{ organisation.name }} </p>
-                <div *ngFor="let unit of organisation?.organisationUnits">
+                <div *ngFor="let unit of organisation?.units">
                   <ul *ngIf="organisation.name !== unit.name" class="nhsuk-list nhsuk-u-margin-left-3 nhsuk-u-margin-bottom-0 border-left-inset-neutral">
                     <li class="d-flex nhsuk-u-margin-0 nhsuk-u-padding-left-3">
                       <span class="app-task-list__task-name"> {{ unit.name }} </span>
