+++ src/modules/feature-modules/innovator/services/innovator.service.spec.ts
@@ -250,79 +250,6 @@ describe('FeatureModules/Innovator/InnovatorService', () => {
 
   });
 
-  it('should run getInnovationNeedsAssessment() and return success', () => {
-
-    const responseMock = {
-      id: 'AssessmentId01',
-      innovation: { id: '01', name: 'Innovation 01' },
-      description: 'description',
-      maturityLevel: null,
-      maturityLevelComment: null,
-      hasRegulatoryApprovals: null,
-      hasRegulatoryApprovalsComment: null,
-      hasEvidence: null,
-      hasEvidenceComment: null,
-      hasValidation: null,
-      hasValidationComment: null,
-      hasProposition: null,
-      hasPropositionComment: null,
-      hasCompetitionKnowledge: null,
-      hasCompetitionKnowledgeComment: null,
-      hasImplementationPlan: null,
-      hasImplementationPlanComment: null,
-      hasScaleResource: null,
-      hasScaleResourceComment: null,
-      summary: null,
-      organisations: [{ id: 'OrgId01', name: 'Organisation Name', acronym: 'ORG', organisationUnits: [{ id: 'OrgId01', name: 'Organisation Unit Name', acronym: 'Unit' }] }],
-      assignToName: '',
-      finishedAt: null,
-      updatedBy: null,
-      updatedAt: null,
-      createdAt: null,
-      createdBy: null
-    };
-
-    const expected = {
-      innovation: responseMock.innovation,
-      assessment: {
-        description: responseMock.description,
-        maturityLevel: responseMock.maturityLevel,
-        maturityLevelComment: responseMock.maturityLevelComment,
-        hasRegulatoryApprovals: responseMock.hasRegulatoryApprovals,
-        hasRegulatoryApprovalsComment: responseMock.hasRegulatoryApprovalsComment,
-        hasEvidence: responseMock.hasEvidence,
-        hasEvidenceComment: responseMock.hasEvidenceComment,
-        hasValidation: responseMock.hasValidation,
-        hasValidationComment: responseMock.hasValidationComment,
-        hasProposition: responseMock.hasProposition,
-        hasPropositionComment: responseMock.hasPropositionComment,
-        hasCompetitionKnowledge: responseMock.hasCompetitionKnowledge,
-        hasCompetitionKnowledgeComment: responseMock.hasCompetitionKnowledgeComment,
-        hasImplementationPlan: responseMock.hasImplementationPlan,
-        hasImplementationPlanComment: responseMock.hasImplementationPlanComment,
-        hasScaleResource: responseMock.hasScaleResource,
-        hasScaleResourceComment: responseMock.hasScaleResourceComment,
-        summary: responseMock.summary,
-        organisations: responseMock.organisations,
-        assignToName: responseMock.assignToName,
-        finishedAt: responseMock.finishedAt,
-        updatedBy: null,
-        updatedAt: null,
-        createdAt: null,
-        createdBy: null
-      }
-    };
-
-    let response: any = null;
-    service.getInnovationNeedsAssessment('Inno01', 'AssessmentId01').subscribe({ next: success => response = success, error: error => response = error });
-
-    const httpRequest = httpMock.expectOne(`${envVariablesStore.API_URL}/innovators/UserId01/innovations/Inno01/assessments/AssessmentId01`);
-    httpRequest.flush(responseMock);
-    expect(httpRequest.request.method).toBe('GET');
-    expect(response).toEqual(expected);
-
-  });
-
   it('should run getInnovationTransfers() with NO parameter and return success', () => {
 
     const responseMock = [{
