+++ src/modules/feature-modules/innovator/services/innovator.service.ts
@@ -42,47 +42,6 @@ export type getInnovationActionsListEndpointOutDTO = {
   closedActions: (getInnovationActionsListEndpointInDTO & { name: string })[];
 };
 
-export type getInnovationNeedsAssessmentEndpointInDTO = {
-  id: string;
-  innovation: { id: string; name: string; };
-  description: null | string;
-  maturityLevel: null | string;
-  maturityLevelComment: null | string;
-  hasRegulatoryApprovals: null | string;
-  hasRegulatoryApprovalsComment: null | string;
-  hasEvidence: null | string;
-  hasEvidenceComment: null | string;
-  hasValidation: null | string;
-  hasValidationComment: null | string;
-  hasProposition: null | string;
-  hasPropositionComment: null | string;
-  hasCompetitionKnowledge: null | string;
-  hasCompetitionKnowledgeComment: null | string;
-  hasImplementationPlan: null | string;
-  hasImplementationPlanComment: null | string;
-  hasScaleResource: null | string;
-  hasScaleResourceComment: null | string;
-  summary: null | string;
-  organisations: { id: string; name: string; acronym: null | string; organisationUnits: { id: string; name: string; acronym: string; }[]; }[];
-  assignToName: string;
-  finishedAt: null | string;
-  updatedBy: null | string;
-  updatedAt: null | string;
-  createdAt: null | string;
-  createdBy: null | string;
-};
-export type getInnovationNeedsAssessmentEndpointOutDTO = {
-  innovation: { id: string; name: string; };
-  assessment: Omit<getInnovationNeedsAssessmentEndpointInDTO, 'id' | 'innovation' | 'organisations'> & {
-    organisations: {
-      id: string;
-      name: string;
-      acronym: null | string;
-      organisationUnits: { id: string; name: string; acronym: string; }[];
-    }[];
-  }
-};
-
 export type GetSupportLogListInDTO = {
   id: string;
   type: 'ACCESSOR_SUGGESTION' | 'STATUS_UPDATE',
@@ -178,45 +137,6 @@ export class InnovatorService extends CoreService {
       take(1),
       map(response => response)
     );
-  }
-
-  getInnovationNeedsAssessment(innovationId: string, assessmentId: string): Observable<getInnovationNeedsAssessmentEndpointOutDTO> {
-
-    const url = new UrlModel(this.API_URL).addPath('innovators/:userId/innovations/:innovationId/assessments/:assessmentId').setPathParams({ userId: this.stores.authentication.getUserId(), innovationId, assessmentId });
-    return this.http.get<getInnovationNeedsAssessmentEndpointInDTO>(url.buildUrl()).pipe(
-      take(1),
-      map(response => ({
-        innovation: response.innovation,
-        assessment: {
-          description: response.description,
-          maturityLevel: response.maturityLevel,
-          maturityLevelComment: response.maturityLevelComment,
-          hasRegulatoryApprovals: response.hasRegulatoryApprovals,
-          hasRegulatoryApprovalsComment: response.hasRegulatoryApprovalsComment,
-          hasEvidence: response.hasEvidence,
-          hasEvidenceComment: response.hasEvidenceComment,
-          hasValidation: response.hasValidation,
-          hasValidationComment: response.hasValidationComment,
-          hasProposition: response.hasProposition,
-          hasPropositionComment: response.hasPropositionComment,
-          hasCompetitionKnowledge: response.hasCompetitionKnowledge,
-          hasCompetitionKnowledgeComment: response.hasCompetitionKnowledgeComment,
-          hasImplementationPlan: response.hasImplementationPlan,
-          hasImplementationPlanComment: response.hasImplementationPlanComment,
-          hasScaleResource: response.hasScaleResource,
-          hasScaleResourceComment: response.hasScaleResourceComment,
-          summary: response.summary,
-          finishedAt: response.finishedAt,
-          assignToName: response.assignToName,
-          updatedAt: response.updatedAt,
-          updatedBy: response.updatedBy,
-          createdAt: response.createdAt,
-          createdBy: response.createdBy,
-          organisations: response.organisations,
-        }
-      }))
-    );
-
   }
 
   getSupportLogList(innovationId: string): Observable<GetSupportLogListOutDTO[]> {
